+++ benchmarks/internal/thread-worker.mjs
@@ -2,11 +2,11 @@ import { isMainThread } from 'node:worker_threads'
 
 import { ThreadWorker } from '../../lib/index.mjs'
 import { TaskFunctions } from '../benchmarks-types.cjs'
-import { executeTaskFunction } from '../benchmarks-utils.cjs'
+import { executeTaskFunction } from '../benchmarks-utils.mjs'
 
 const taskFunction = data => {
   data = data || {}
-  data.function = data.function || TaskFunctions.jsonIntegerSerialization
+  data.function = data.function || TaskFunctions.factorial
   data.debug = data.debug || false
   const res = executeTaskFunction(data)
   data.debug === true &&